# openapi.yaml
openapi: 3.0.3
info:
  title: dschool Student Attendance API
  description: 'RESTful Web API สำหรับตรวจสอบข้อมูลการ **ขาด ลา มาสาย เข้าเรียน**
    ของนักเรียน

    แปลงข้อมูลจากระบบ dschool (PHP Web Application) ให้กลายเป็น JSON Web API มาตรฐาน

    '
  version: 1.0.0
  contact:
    name: API Support
servers:
- url: http://localhost:8080
  description: Local Development Server
paths:
...
  /api/attendance/overview:
    get:
      summary: School Overview Attendance
      description: สรุปภาพรวมการลงเวลาทั้งโรงเรียน ยอดรวมแยกรายระดับชั้น (ม.1 - ม.6)
        และแยกรายห้อง
      operationId: getOverviewAttendance
      tags:
      - Attendance
      parameters:
      - name: date
        in: query
        description: วันที่ที่ต้องการตรวจสอบ (รูปแบบ YYYY-MM-DD) ค่าเริ่มต้นคือวันนี้
        required: false
        schema:
          type: string
          example: '2026-09-08'
      responses:
        '200':
          description: ดึงข้อมูลภาพรวมสำเร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewAttendanceResponse'
        '400':
          description: รูปแบบวันที่ไม่ถูกต้อง
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/students/search:
    get:
      summary: Search Students
      description: ค้นหานักเรียนด้วยชื่อ นามสกุล หรือรหัสนักเรียน เพื่อรับรหัสภายใน
        (sd_no) และรูปถ่าย
      operationId: searchStudents
      tags:
      - Students
      parameters:
      - name: q
        in: query
        description: คำค้นหา (ชื่อ นามสกุล หรือรหัสนักเรียน)
        required: true
        schema:
          type: string
          example: กิตติพิชญ์
      responses:
        '200':
          description: ค้นหาสำเร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentSearchResponse'
        '400':
          description: ไม่ได้ระบุคำค้นหา หรือคำค้นหายาวเกินขนาด
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/keys:
    get:
      summary: List All API Keys (Admin)
      description: ดึงรายการ API Key ทั้งหมดในระบบ พร้อมสถานะการใช้งาน วันที่สร้าง
        และเวลาที่ใช้งานล่าสุด
      operationId: listApiKeys
      tags:
      - Admin - Key Management
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      responses:
        '200':
          description: ดึงรายการคีย์สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/KeyInfo'
        '403':
          description: ไม่มีสิทธิ์ Admin
    post:
      summary: Create / Issue New API Key (Admin)
      description: สร้างหรือออก API Key ใหม่ในระบบ กำหนดชื่อ Role และสามารถกำหนดรหัสเองหรือให้ระบบสุ่มรหัสให้อัตโนมัติ
      operationId: createApiKey
      tags:
      - Admin - Key Management
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              properties:
                name:
                  type: string
                  example: LINE Bot Service
                key:
                  type: string
                  example: dsk_live_custom_key_123
                role:
                  type: string
                  enum:
                  - client
                  - admin
                  default: client
      responses:
        '201':
          description: สร้าง API Key สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/KeyInfo'
        '400':
          description: ข้อมูลนำเข้าไม่ถูกต้อง หรือคีย์ซ้ำ
  /api/admin/keys/revoke:
    post:
      summary: Revoke API Key (Admin)
      description: เพิกถอนสิทธิ์การใช้งาน API Key ทำให้ไม่สามารถนำไปเรียกใช้งาน API
        ได้อีกต่อไป (ตอบกลับ 401 ทันที)
      operationId: revokeApiKey
      tags:
      - Admin - Key Management
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - key
              properties:
                key:
                  type: string
                  example: dsk_live_04abd0f4792ed29198ab5f23fa3a9428
      responses:
        '200':
          description: เพิกถอนสิทธิ์คีย์สำเร็จ
        '404':
          description: ไม่พบคีย์ในระบบ
        '403':
          description: ไม่อนุญาตให้เพิกถอน Master Admin Key
  /api/admin/keys/activate:
    post:
      summary: Re-activate API Key (Admin)
      description: เปิดใช้งาน API Key ที่เคยถูกเพิกถอนให้กลับมาใช้งานได้ใหม่อีกครั้ง
      operationId: activateApiKey
      tags:
      - Admin - Key Management
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - key
              properties:
                key:
                  type: string
                  example: dsk_live_04abd0f4792ed29198ab5f23fa3a9428
      responses:
        '200':
          description: เปิดใช้งานคีย์สำเร็จ
        '404':
          description: ไม่พบคีย์ในระบบ
components:
  schemas:
    BaseResponse:
      type: object
      required:
      - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
        error:
          type: string
          example: null
    ErrorResponse:
      type: object
      required:
      - success
      - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Invalid parameter
...
    AttendanceCounts:
      type: object
      properties:
        total:
          type: integer
          example: 3534
        present:
          type: integer
          example: 3340
        off_campus:
          type: integer
          example: 60
        unrecorded:
          type: integer
          example: 55
        sick_leave:
          type: integer
          example: 49
        personal_leave:
          type: integer
          example: 20
        absent:
          type: integer
          example: 10
    GradeLevelAttendance:
      allOf:
      - type: object
        properties:
          grade:
            type: string
            example: ม.1
      - $ref: '#/components/schemas/AttendanceCounts'
    ClassroomAttendance:
      allOf:
      - type: object
        properties:
          classroom:
            type: string
            example: ม.101
      - $ref: '#/components/schemas/AttendanceCounts'
    OverviewAttendanceData:
      type: object
      properties:
        date:
          type: string
          example: 8 ก.ย. 2569
        summary:
          $ref: '#/components/schemas/AttendanceCounts'
        grade_levels:
          type: array
          items:
            $ref: '#/components/schemas/GradeLevelAttendance'
        classrooms:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/ClassroomAttendance'
    OverviewAttendanceResponse:
      allOf:
      - $ref: '#/components/schemas/BaseResponse'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/OverviewAttendanceData'
...
    StudentSearchResult:
      type: object
      properties:
        false:
          type: integer
          example: 1
        sd_no:
          type: string
          example: '18948'
        name:
          type: string
          example: กิตติพิชญ์ ปัชชามูล
        photo_url:
          type: string
          example: https://dschool-l1.gp-education.com/dschool_app_v2020/card_template/img_proxy.php?...
    StudentSearchResponse:
      allOf:
      - $ref: '#/components/schemas/BaseResponse'
      - type: object
        properties:
          data:
            type: array
            items:
              $ref: '#/components/schemas/StudentSearchResult'
...
    KeyInfo:
      type: object
      properties:
        key:
          type: string
          example: dsk_live_04abd0f4792ed29198ab5f23fa3a9428
        name:
          type: string
          example: LINE Bot Service
        is_active:
          type: boolean
          example: true
        role:
          type: string
          example: client
        created_at:
          type: string
          format: date-time
          example: '2026-09-08T19:55:00+07:00'
        revoked_at:
          type: string
          format: date-time
          nullable: true
        last_used_at:
          type: string
          format: date-time
          nullable: true
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication via X-API-Key header
    BearerAuth:
      type: http
      scheme: bearer
      description: 'API Key authentication via Authorization: Bearer <token> header'
    QueryApiKey:
      type: apiKey
      in: query
      name: api_key
      description: API Key authentication via query parameter
security:
- ApiKeyAuth: []
- BearerAuth: []
- QueryApiKey: []